test : mock은 단위테스트/기능 용도로 테스트 (객체 매번 새로 생성)        spring기반 테스트는 실제 DI 확인 및 싱글톤 재사용 확인이 가능하므로        테스트를 mock기반, spring기반으로 나눠서 분리함        jvm 메모리 사용 측정 테스트 - GC는 JVM이 보장된 시점에 수행되지 않음. `runtime.gc()` 는 GC 요청일 뿐, 즉시 실행한다는 보장이 없기 때문.. 단순 참고용
