+++ src/test/java/com/jisutudy/domain/sms/filter/SmsFilterBeforeDIImpl.java
@@ -1,0 +1,76 @@
+package com.jisutudy.domain.sms.filter;
+
+import com.jisutudy.domain.CustSmsConsentType;
+import com.jisutudy.domain.SmsTemplate;
+import com.jisutudy.domain.SmsType;
+import com.jisutudy.domain.sms.Sms;
+import com.jisutudy.domain.sms.SmsResult;
+import com.jisutudy.service.filter.AdvertiseSmsFilter;
+import com.jisutudy.service.filter.CustConsentFilter;
+import com.jisutudy.service.filter.SmsFilter;
+import com.jisutudy.service.filter.TimeSmsFilter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@RequiredArgsConstructor
+@Component
+public class SmsFilterBeforeDIImpl implements SmsFilter {
+
+    private final TimeSmsFilter timeSmsFilter;
+    private final AdvertiseSmsFilter advertiseSmsFilter;
+//    private final CustomerSmsFilter customerSmsFilter;
+
+    @Override
+    public SmsResult filter(Sms sms, CustSmsConsentType type) {
+        // sms 필터링
+        SmsTemplate smsTemplate = sms.getSmsTemplate();
+
+        // 1. 시간
+        if (!timeSmsFilter.isSendable(sms.getSendDt())) {
+            return SmsResult.NOT_SEND_TIME;
+        }
+
+        // 2. 고객동의
+        CustConsentFilter customerSmsFilter = new CustConsentFilter(); // TODO OCP, DIP 위반
+        if (!customerSmsFilter.isSendable(type, smsTemplate.getSmsType())) {
+            return SmsResult.CUST_REJECT;
+        }
+
+        // 3. 광고 개수 제한
+        if (SmsType.ADVERTISING == smsTemplate.getSmsType()) {
+            if (!advertiseSmsFilter.isSendable(sms)) {
+                return SmsResult.AD_COUNT_OVER;
+            }
+        }
+
+        return SmsResult.SUCCESS;
+    }
+
+    @Deprecated
+    @Override
+    public SmsResult filter(Sms sms) {
+//        // sms 필터링
+//        Cust cust = sms.getCust();
+//        SmsTemplate smsTemplate = sms.getSmsTemplate();
+//
+//        // 1. 시간
+//        if (!timeSmsFilter.isSendable(sms.getSendDt())) {
+//            return SmsResult.NOT_SEND_TIME;
+//        }
+//
+//        // 2. 고객동의
+//        CustConsentFilter consentFilter = new CustConsentFilter(); // TODO OCP, DIP 위반
+//        if (!consentFilter.isSendable(cust.getSmsConsentType(), smsTemplate.getSmsType())) {
+//            return SmsResult.CUST_REJECT;
+//        }
+//
+//        // 3. 광고 개수 제한
+//        if (SmsType.ADVERTISING == smsTemplate.getSmsType()) {
+//            if (!advertiseSmsFilter.isSendable(sms)) {
+//                return SmsResult.AD_COUNT_OVER;
+//            }
+//        }
+
+        return SmsResult.SUCCESS;
+    }
+}
