+++ src/test/java/com/jisutudy/domain/sms/filter/TestAdvertiseSmsFilter.java
@@ -1,0 +1,32 @@
+package com.jisutudy.domain.sms.filter;
+
+import com.jisutudy.domain.SmsType;
+import com.jisutudy.domain.sms.Sms;
+import com.jisutudy.repository.SmsSearch;
+import com.jisutudy.service.filter.AdvertiseSmsFilter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.LocalTime;
+
+@RequiredArgsConstructor
+@Component
+public class TestAdvertiseSmsFilter implements AdvertiseSmsFilter {
+
+    @Override
+    public boolean isSendable(Sms sms) {
+        LocalDateTime startDt = LocalDateTime.of(LocalDate.now(), LocalTime.of(0, 0));
+        LocalDateTime endDt = LocalDateTime.of(LocalDate.now().plusDays(1), LocalTime.of(0, 0));
+
+        SmsSearch smsSearch = SmsSearch.builder()
+                .startDt(startDt)
+                .endDt(endDt)
+                .custId(sms.getCustId())
+                .smsType(SmsType.ADVERTISING)
+                .build();
+
+        return true;
+    }
+}
